pyproject.toml
[project]
name = "osap-aiops-agent"
version = "0.1.0"
description = "An AIOps agent for the OSAP platform."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic-settings>=2.14.0",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.30.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = true
